got timer working (done with new-items)
